+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 1.8 (not yet released)
 * Added Latitude and Longitude fields to address actions.
 * New action: 'Add attachment to bulk mail'.
 * Added create a case action.
+* Added action to find contact by external id.
 
 Version 1.7
 -----------